New user folder check in
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@1022 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/tinytcl/tclGlob.c
@@ -359,7 +359,7 @@ DoGlob(interp, dir, rem)
 	} else {
 	    newDir = (char *) ckalloc((unsigned) l2);
 	}
-	sprintf(newDir, "%s%s%.*s", dir, separator, p-rem, rem);
+	sprintf(newDir, "%s%s%.*s", dir, separator, (int)(p-rem), rem);
 	result = DoGlob(interp, newDir, p+1);
 	if (newDir != static1) {
 	    ckfree(newDir);
